chore: some websocket cleanup
turn off zlib compression request when constructing socket since I'm not
using it (as far as I know, not being expert in boost beast).
Also, tell the socket to cancel any outstanding I/O before we close it.
this is an attempt to prevent the 'broken pipe' error I'm getting in
testing.