fix: embedded Python and threads.
First, use a non-GUI backend to avoid warnings about starting a GUI from
not-the-main-thread.  This is OK since we are only going to produce
files for output anyway.  Viewing is done with the graphics viewer of
your choice.

Second, we need to release and acquire the GIL so the code will actually
do stuff.  The code I have works but I'm not sure it's 'correct' or
'optimal'. Maybe another day...