refactoring: extending column up or down part 2
Continue with box based logic.  Just compute where we need to be if
there is to be a change in the chart then see if the new value is there.
We just compute the destination box once and use it where needed.

If it is and it's a reversal, look for 1-step-back otherwise go for it.
Use computed destination box from the code above.