feature: put a limit on number of boxes
It is possible to end up with a tiny boxsize when using ATR and a
fraction for chart box size.  When that happens, and there is a lot of
data, processing slows to a crawl. To avoid that, set a maximum number
of boxes and throw an exception when that is exceeded.

For now, the maximum is hard coded and shared by all boxes instances.